+++ aiohttp_things/__init__.py
@@ -1,4 +1,4 @@
 from aiohttp_things import views
 
-__version__ = '0.2.0'
+__version__ = '0.6.0'
 __all__ = ['views']